Initial Setup
Created the game onto Github, separating the style and scripts into it's own folders.
Index is setup to just have "Welcome!" and a start button in the center.
Scripts are divided into 2 files (enterGame and Game).
(a) Game is the initial version of the game. The reset function may have to be revised to return to the main menu instead.
(b) enterGame are new codes written to change the UI once it's hit. These codes contains the html that would have been on the index file of the game.
The ReadMe file is just a copy and paste of the 2 questions.